requirements.txt
# Chromatic Mood of Fashion Eras - Requirements
# Python 3.8+ recommended

# Core Data Science
numpy>=1.21.0
pandas>=1.3.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.5.0

# Machine Learning
scikit-learn>=1.0.0

# Color Science
colormath>=3.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Progress Tracking
tqdm>=4.62.0

# Web Requests
requests>=2.26.0

# Jupyter (optional, for notebook use)
jupyter>=1.0.0
jupyterlab>=3.0.0
ipywidgets>=7.6.0



